+++ oinkbrew/Helper.cpp
@@ -170,3 +170,23 @@ void Helper::setBytes(uint8_t* data, const char* s, uint8_t len) {
 		*data++ = d;
 	}
 }
+
+/*******************************************************************************
+ * Function Name  : matchAddress
+ * Description    : tests if two DeviceAddress are the same
+ * Input          : hw_adress to check and the number of bytes
+ * Output         : true if a match, false if not matching
+ * Return         :
+ ******************************************************************************/
+bool Helper::matchAddress(uint8_t* detected, uint8_t* configured, uint8_t count) {
+
+	if (!configured[0])
+		return true;
+
+	while (count-- > 0) {
+		if (detected[count] != configured[count])
+			return false;
+	}
+
+	return true;
+}