+++ oinkbrew/Helper.h
@@ -43,6 +43,8 @@ public:
 
     static void getBytes(const uint8_t* data, uint8_t len, char* buf);
     static void setBytes(uint8_t* data, const char* s, uint8_t len);
+
+    static bool matchAddress(uint8_t* detected, uint8_t* configured, uint8_t count);
 };
 
 extern Helper helper;