+++ oinkbrew/TcpListener.h
@@ -54,7 +54,7 @@ private:
     
     static void processSparkInfo(const char * key, const char * val, void* pv);
     static void setDeviceMode(const char * key, const char * val, void* pv);
-    static void receiveToggleRequest(const char * key, const char * val, void* pv);
+    static void receiveDeviceRequest(const char * key, const char * val, void* pv);
     
 };