+++ screens/info_screen.h
@@ -37,7 +37,6 @@ D4D_EXTERN_SCREEN(screen_info);
 D4D_EXTERN_OBJECT(scrInfo_valName);
 D4D_EXTERN_OBJECT(scrInfo_valMode);
 D4D_EXTERN_OBJECT(scrInfo_valId);
-D4D_EXTERN_OBJECT(scrInfo_valConfig);
 D4D_EXTERN_OBJECT(scrInfo_valTemp);
 D4D_EXTERN_OBJECT(scrInfo_valFirmware);
 D4D_EXTERN_OBJECT(scrInfo_valIp);