+++ screens/startup_screen.c
@@ -35,9 +35,9 @@ D4D_CHAR str_version[15] = "";
         
 D4D_DECLARE_STD_PICTURE(scrStartup_bmpLogo, 120, 10, 80, 68, &bmp_oink_logo);
 
-D4D_DECLARE_STD_LABEL_CENTER(scrStartup_message, str_message, 40, 140, 240, 20, FONT_ARIAL_11, NULL);
+D4D_DECLARE_STD_LABEL_CENTER(scrStartup_message, str_message, 40, 140, 240, 20, FONT_ARIAL_11, D4D_COLOR_BLACK, D4D_COLOR_WHITE);
 D4D_DECLARE_STD_LABEL(scrStartup_version, str_version, 5, 220, 80, 20, FONT_ARIAL_11);
-D4D_DECLARE_STD_LABEL_RIGHT(scrStartup_copyright, "(c) 2015 by Oink Brew", 170, 220, 145, 20, FONT_ARIAL_11, NULL);
+D4D_DECLARE_STD_LABEL_RIGHT(scrStartup_copyright, "(c) 2015 by Oink Brew", 170, 220, 145, 20, FONT_ARIAL_11, D4D_COLOR_BLACK, D4D_COLOR_WHITE);
 
 
 D4D_DECLARE_STD_SCREEN_BEGIN(screen_startup, ScreenStartup_)